fixing test
