+++ src/test/java/com/vmware/vip/i18n/LocaleTest.java
@@ -6,14 +6,9 @@ package com.vmware.vip.i18n;
 
 import com.vmware.vipclient.i18n.I18nFactory;
 import com.vmware.vipclient.i18n.VIPCfg;
-import com.vmware.vipclient.i18n.base.DataSourceEnum;
 import com.vmware.vipclient.i18n.base.cache.*;
 import com.vmware.vipclient.i18n.base.instances.LocaleMessage;
 import com.vmware.vipclient.i18n.exceptions.VIPClientInitException;
-import com.vmware.vipclient.i18n.messages.dto.BaseDTO;
-import com.vmware.vipclient.i18n.messages.dto.MessagesDTO;
-import com.vmware.vipclient.i18n.messages.service.CacheService;
-import com.vmware.vipclient.i18n.messages.service.ProductService;
 import com.vmware.vipclient.i18n.util.LocaleUtility;
 import org.json.simple.parser.ParseException;
 import org.junit.Assert;
@@ -22,7 +17,6 @@ import org.junit.Test;
 
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.LinkedList;
 import java.util.List;
 import java.util.Locale;
 import java.util.Map;
@@ -117,49 +111,6 @@ public class LocaleTest extends BaseTestClass {
         //test invalid locale, fallback to default locale
         Map<String, String> aaResp = localeI18n.getDisplayLanguagesList("aa");
         Assert.assertNotNull(aaResp);
-    }
-    
-    @Test
-    public void testGetSupportedLocalesOfflineBundles() throws ParseException {
-    	//Enable offline mode
-    	String offlineResourcesBaseUrlOrig = gc.getOfflineResourcesBaseUrl();
-    	gc.setOfflineResourcesBaseUrl("offlineBundles/");
-    	List<DataSourceEnum> msgOriginsQueueOrig = gc.getMsgOriginsQueue();
-    	gc.setMsgOriginsQueue(new LinkedList<DataSourceEnum>(Arrays.asList(DataSourceEnum.Bundle)));
-
-    	// List of supported locales shall be determined from available offline bundle files.
-        Map<String, String> resp = localeI18n.getDisplayLanguagesList("fil");
-        Assert.assertTrue(resp.containsKey("fil"));
-
-        MessagesDTO msgsDTO = new MessagesDTO("JAVA", "fil-PH", gc.getProductName(), gc.getVersion());
-        CacheService cs = new CacheService(msgsDTO);
-
-        Cache c = gc.createTranslationCache(MessageCache.class);
-        TranslationCacheManager.cleanCache(c);
-        gc.createTranslationCache(MessageCache.class);
-
-        BaseDTO dto = new BaseDTO();
-        dto.setVersion(gc.getVersion());
-        dto.setProductID(gc.getProductName());
-        ProductService ps = new ProductService(dto);
-        ps.getAllComponentTranslation();
-
-        // Locale "fil" MessageCacheItem exists in cache
-    	List<Locale> localesOfCachedMessages = cs.getLocalesOfCachedMsgs();
-        Assert.assertTrue(localesOfCachedMessages.contains(Locale.forLanguageTag("fil")));
-
-        // Locale "fil-PH" MessageCacheItem exists in cache
-        Assert.assertFalse(localesOfCachedMessages.contains(Locale.forLanguageTag("fil-PH")));
-
-        // cs.getCacheOfComponent() for fil-PH returns MessageCacheItem of fil
-        MessageCacheItem filPHCacheItem = cs.getCacheOfComponent();
-        Assert.assertTrue(filPHCacheItem.getLocale().equals("fil"));
-
-        localeI18n.getDisplayLanguagesList("invalid_locale");// get data from cache
-        
-        // Disable offline mode off for next tests.
-        gc.setOfflineResourcesBaseUrl(offlineResourcesBaseUrlOrig);
-    	gc.setMsgOriginsQueue(msgOriginsQueueOrig);
     }
 
     @Test
